@@ -7,84 +7,70 @@
     "email": "marko.j.niemi@gmail.com"
   },
   "license": "MIT",
-  "repository": "github:markoniemi/oidc-demo",
-  "engines" : {
-    "node" : "^18.18.0",
-    "npm": "8.19.4"
+  "repository": "github:markoniemi/spring-boot-react-demo",
+  "engines": {
+    "node": "22.12.0",
+    "npm": "11.4.2"
   },
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "6.5.2",
-    "@fortawesome/free-solid-svg-icons": "6.5.2",
+    "@fortawesome/fontawesome-svg-core": "6.7.2",
+    "@fortawesome/free-solid-svg-icons": "6.7.2",
     "@fortawesome/react-fontawesome": "0.2.2",
-    "bootstrap": "5.3.3",
-    "css-loader": "7.1.2",
-    "file-loader": "6.2.0",
+    "bootstrap": "5.3.7",
     "formik": "2.4.6",
     "isomorphic-fetch": "3.0.0",
-    "loglevel": "1.9.1",
+    "loglevel": "1.9.2",
     "oidc-client-ts": "3.0.1",
-    "react": "18.3.1",
-    "react-bootstrap": "2.10.2",
-    "react-dom": "18.3.1",
-    "react-intl": "6.6.8",
+    "react": "19.1.0",
+    "react-bootstrap": "2.10.10",
+    "react-dom": "19.1.0",
+    "react-intl": "7.1.11",
     "react-oidc-context": "3.1.0",
-    "react-router-dom": "6.23.1",
-    "style-loader": "4.0.0",
-    "yup": "1.4.0"
+    "react-router": "7.6.2",
+    "yup": "1.6.1"
   },
   "scripts": {
     "clean": "shx rm -rf build & shx rm -rf dist",
-    "compile": "tsc && set NODE_OPTIONS=--openssl-legacy-provider && webpack --mode production --config dist/webpack.prod.js",
-    "start": "tsc && set NODE_OPTIONS=--openssl-legacy-provider && webpack serve --mode development --config dist/webpack.config.js",
-    "test": "jest --coverage --config jest.config.ts",
+    "test": "vitest --coverage",
     "coveralls": "cat ./reports/coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js",
     "sonarqube": "./node_modules/.bin/sonar-scanner -Dsonar.login=$SONAR_CLOUD_TOKEN",
-    "dependency-tree": "npm ls -a > dependencyTree.txt",
-    "update": "npx dependency-time-machine --update --install --auto",
-    "auditFix": "npm audit fix"
+    "dependency-tree": "npm ls -a > reports/dependencyTree.txt",
+    "update": "npx npm-check-updates -u",
+    "auditFix": "npm audit fix",
+    "dev": "vite --debug",
+    "build": "vite build",
+    "compile": "tsc --noEmit"
   },
   "devDependencies": {
-    "@testing-library/react": "14.3.1",
-    "@types/chai": "4.3.16",
-    "@types/dotenv-webpack": "7.0.7",
-    "@types/fetch-mock": "7.3.8",
+    "@testing-library/react": "16.3.0",
+    "@testing-library/user-event": "14.6.1",
+    "@tsconfig/vite-react": "6.3.6",
+    "@types/fetch-mock": "9.2.2",
     "@types/isomorphic-fetch": "0.0.39",
-    "@types/jest": "29.5.12",
-    "@types/node": "18.15.3",
-    "@types/react": "18.3.3",
-    "@types/react-dom": "18.3.0",
-    "@types/sinon": "17.0.3",
-    "@types/webpack-env": "1.16.2",
-    "@typescript-eslint/eslint-plugin": "7.12.0",
-    "@typescript-eslint/parser": "7.12.0",
-    "chai": "4.4.1",
-    "copy-webpack-plugin": "12.0.2",
-    "dotenv": "16.4.5",
-    "dotenv-webpack": "8.1.0",
-    "es7-sleep": "1.0.0",
-    "eslint": "8.57.0",
-    "eslint-config-prettier": "8.10.0",
-    "eslint-plugin-prettier": "5.1.3",
-    "eslint-plugin-react": "7.34.2",
-    "fetch-mock": "5.13.1",
-    "history": "4.10.1",
-    "jest": "29.7.0",
-    "jest-environment-jsdom": "29.7.0",
-    "jest-junit": "16.0.0",
-    "jest-localstorage-mock": "2.4.26",
-    "jest-sonar-reporter": "2.0.0",
-    "prettier": "3.3.1",
+    "@types/jsdom": "21.1.7",
+    "@types/node": "24.0.4",
+    "@types/react": "19.1.8",
+    "@types/react-dom": "19.1.6",
+    "@typescript-eslint/eslint-plugin": "8.35.0",
+    "@typescript-eslint/parser": "8.35.0",
+    "@vitejs/plugin-react": "4.6.0",
+    "@vitejs/plugin-react-refresh": "1.3.6",
+    "@vitest/coverage-v8": "3.2.4",
+    "@vitest/ui": "3.2.4",
+    "dotenv": "16.5.0",
+    "eslint": "9.29.0",
+    "eslint-config-prettier": "10.1.5",
+    "eslint-plugin-prettier": "5.5.1",
+    "eslint-plugin-react": "7.37.5",
+    "fetch-mock": "12.5.3",
+    "history": "5.3.0",
+    "jsdom": "26.1.0",
+    "prettier": "3.6.2",
     "raf": "3.4.1",
-    "shx": "0.3.4",
-    "sinon": "18.0.0",
-    "source-map-loader": "5.0.0",
-    "ts-node": "10.9.2",
-    "ts-jest": "29.1.4",
-    "ts-loader": "9.5.1",
-    "typescript": "5.4.5",
-    "webpack": "5.91.0",
-    "webpack-cli": "5.1.4",
-    "webpack-dev-server": "4.3.0"
+    "shx": "0.4.0",
+    "typescript": "5.8.3",
+    "vite": "7.0.0",
+    "vitest": "3.2.4"
   },
   "proxy": "http://localhost:8080",
   "browserslist": [
