@@ -1,4 +1,4 @@
-import {OidcClientSettings, User} from "oidc-client-ts";
+import {type OidcClientSettings, User} from "oidc-client-ts";
 
 export default class OidcService {
     public static getUser() {
@@ -16,11 +16,12 @@ export default class OidcService {
         return headers;
     }
 
-    public static getToken(): string {
+    public static getToken(): string | undefined {
         return OidcService.getUser()?.id_token;
     }
 }
 // TODO move this inside class
+// https://console.cloud.google.com/apis/credentials?project=oidc-demo-426909
 export const oidcConfig: OidcClientSettings = {
     authority: "https://accounts.google.com",
     client_id: "977467276012-1pkp49dcqgj8k0f3dprbl53avt8sgcep.apps.googleusercontent.com",
