@@ -1,9 +1,9 @@
 import React from "react";
-import { AuthContextProps, withAuth } from "react-oidc-context";
-import Empty from "../domain/Empty";
+import {type AuthContextProps, withAuth } from "react-oidc-context";
+import {type Empty} from "../domain/Empty";
 import UserContainer from "./UsersContainer";
 import EditUser from "./EditUser";
-import { BrowserRouter, Navigate, Route, Routes } from "react-router-dom";
+import {BrowserRouter, Navigate, Route, Routes} from "react-router";
 import { IntlProvider } from "react-intl";
 import i18nConfig from "../messages/messages";
 import LoginForm from "./LoginForm";
