@@ -1,9 +1,9 @@
-import React from "react";
+import * as React from "react";
 import TimeServiceImpl from "../api/TimeServiceImpl";
-import { Button } from "react-bootstrap";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import {Button} from "react-bootstrap";
+import {FontAwesomeIcon} from "@fortawesome/react-fontawesome";
 import * as Icons from "@fortawesome/free-solid-svg-icons";
-import Empty from "../domain/Empty";
+import type {Empty} from "../domain/Empty";
 
 export interface TimeState {
     message: string;
@@ -14,7 +14,7 @@ export default class Time extends React.Component<Empty, TimeState> {
 
     constructor(props: Empty) {
         super(props);
-        this.state = { message: "" };
+        this.state = {message: ""};
         this.fetchMessage = this.fetchMessage.bind(this);
     }
 
@@ -24,14 +24,14 @@ export default class Time extends React.Component<Empty, TimeState> {
 
     public async fetchMessage(): Promise<void> {
         const message = await this.helloService.getTime();
-        this.setState({ message: message });
+        this.setState({message: message});
     }
 
     public override render(): React.ReactNode {
         return (
             <>
                 <Button id="fetchMessage" size="sm" onClick={this.fetchMessage}>
-                    <FontAwesomeIcon icon={Icons.faClock} />
+                    <FontAwesomeIcon icon={Icons.faClock}/>
                 </Button>
                 <span id="message">{this.state.message}</span>
             </>
