@@ -4,12 +4,16 @@ const i18nConfig = {
         user: "User",
         users: "Users",
         id: "Id",
+        id_label: "Id",
         login: "Login",
-        "login.google": "Login with Google",
         username: "Username",
+        username_label: "Username",
         email: "Email",
+        email_label: "Email",
         password: "Password",
+        password_label: "Password",
         role: "Role",
+        role_label: "Role",
         "role.ROLE_USER": "User",
         "role.ROLE_ADMIN": "Admin",
         "role.undefined": " ",
