@@ -1,5 +1,5 @@
-import { assert } from "chai";
 import Jwt from "../../src/api/Jwt";
+import {assert, describe, test} from "vitest";
 
 describe("Jwt", () => {
     test("isAuthenticated returns false if sessionStorage does not have token", () => {
