@@ -1,19 +1,20 @@
-import { assert } from "chai";
 import * as dotenv from "dotenv";
 import "isomorphic-fetch";
-import UserService from "../../src/api/UserService";
+import type UserService from "../../src/api/UserService";
 import UserServiceImpl from "../../src/api/UserServiceImpl";
 import fetchMock from "fetch-mock";
-import { user1, users } from "../users";
+import {user1, users} from "../users";
 import User from "../../src/domain/User";
+import {afterEach, assert, beforeEach, describe, expect, test} from "vitest";
 
 const userService: UserService = new UserServiceImpl();
 describe("UserService", () => {
     beforeEach(() => {
-        dotenv.config({ path: "config/development.env" });
+        dotenv.config({path: ".env"});
+        fetchMock.mockGlobal();
     });
     afterEach(() => {
-        fetchMock.restore();
+        fetchMock.hardReset();
     });
     test("fetchUsers", async () => {
         fetchMock.getOnce("/api/rest/users/", users);
@@ -58,7 +59,7 @@ describe("UserService", () => {
     });
     test("delete", async () => {
         fetchMock.deleteOnce("/api/rest/users/2", 200);
-        const savedUser = await userService.delete(2);
+        await userService.delete(2);
     });
     test("delete fails", async () => {
         fetchMock.deleteOnce("/api/rest/users/2", 500);
